CVPR-2026/paper-2205-ChordEdit/requirement.txt
torch==2.5.0+cu124
torchaudio==2.5.0+cu124
torchvision==0.20.0+cu124
torchmetrics==1.8.2
torch-fidelity==0.3.0
numpy
matplotlib
seaborn
Pillow
pyyaml
datasets
transformers
diffusers
gradio
